requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#

-e file:.
    # via -r requirements.in
astor==0.8.1
    # via
    #   flake8-qgis
    #   flake8-simplify
atomicwrites==1.4.0
    # via
    #   -r requirements.in
    #   pytest
attrs==22.1.0
    # via
    #   flake8-bugbear
    #   pytest
black==22.3.0
    # via -r requirements.in
certifi==2022.9.24
    # via requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via black
colorama==0.4.5
    # via
    #   click
    #   pytest
coverage[toml]==6.5.0
    # via pytest-cov
distlib==0.3.6
    # via virtualenv
filelock==3.8.0
    # via virtualenv
flake8==3.9.1
    # via
    #   -r requirements.in
    #   flake8-annotations
    #   flake8-bugbear
    #   flake8-comprehensions
    #   flake8-no-pep420
    #   flake8-polyfill
    #   flake8-print
    #   flake8-qgis
    #   flake8-simplify
    #   flake8-spellcheck
    #   flake8-tidy-imports
flake8-annotations==2.6.2
    # via -r requirements.in
flake8-bugbear==21.4.3
    # via -r requirements.in
flake8-comprehensions==3.6.1
    # via -r requirements.in
flake8-no-pep420==1.1.1
    # via -r requirements.in
flake8-pie==0.14.0
    # via -r requirements.in
flake8-plugin-utils==1.3.2
    # via flake8-pytest-style
flake8-polyfill==1.0.2
    # via pep8-naming
flake8-print==4.0.0
    # via -r requirements.in
flake8-pytest-style==1.5.0
    # via -r requirements.in
flake8-qgis==1.0.0
    # via -r requirements.in
flake8-simplify==0.14.1
    # via -r requirements.in
flake8-spellcheck==0.24.0
    # via -r requirements.in
flake8-tidy-imports==4.4.1
    # via -r requirements.in
identify==2.5.6
    # via pre-commit
idna==3.4
    # via requests
importlib-metadata==5.0.0
    # via qgis-plugin-dev-tools
iniconfig==1.1.1
    # via pytest
isort==5.8.0
    # via -r requirements.in
mccabe==0.6.1
    # via flake8
mypy==0.931
    # via -r requirements.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
nodeenv==1.7.0
    # via pre-commit
packaging==21.3
    # via
    #   pytest
    #   qgis-plugin-dev-tools
pathspec==0.10.1
    # via black
pep8-naming==0.11.1
    # via -r requirements.in
platformdirs==2.5.2
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.15.0
    # via -r requirements.in
py==1.11.0
    # via pytest
pycodestyle==2.7.0
    # via
    #   flake8
    #   flake8-print
pyflakes==2.3.1
    # via flake8
pyparsing==3.0.9
    # via packaging
pytest==6.2.5
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-dotenv
    #   pytest-mock
    #   pytest-order
    #   pytest-qgis
    #   pytest-qt
    #   pytest-timeout
pytest-cov==2.12.0
    # via -r requirements.in
pytest-dotenv==0.5.2
    # via -r requirements.in
pytest-mock==3.7.0
    # via -r requirements.in
pytest-order==1.0.0
    # via -r requirements.in
pytest-qgis==1.3.0
    # via -r requirements.in
pytest-qt==3.3.0
    # via -r requirements.in
pytest-timeout==1.4.2
    # via -r requirements.in
python-dotenv==0.21.0
    # via
    #   pytest-dotenv
    #   qgis-plugin-dev-tools
pyyaml==6.0
    # via pre-commit
qgis-plugin-dev-tools==0.5.0
    # via -r requirements.in
qgis-plugin-tools==0.2.0
    # via quality-result-gui
requests==2.28.1
    # via qgis-plugin-dev-tools
six==1.16.0
    # via flake8-print
toml==0.10.2
    # via
    #   pre-commit
    #   pytest
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   mypy
    #   qgis-plugin-dev-tools
typing-extensions==4.4.0
    # via
    #   black
    #   flake8-pie
    #   mypy
urllib3==1.26.12
    # via requests
virtualenv==20.16.5
    # via pre-commit
zipp==3.10.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
